+++ piet/src/null_renderer.rs
@@ -157,7 +157,7 @@ impl TextLayout for NullTextLayout {
         42.0
     }
 
-    fn hit_test_point(&self, _x: f32, _y: f32) -> HitTestPoint {
+    fn hit_test_point(&self, _x: f64, _y: f64) -> HitTestPoint {
         HitTestPoint::default()
     }
